net/ieee802154: add supported PHY Kconfig symbols
These symbols are meant to be selected by actual IEEE 802.15.4
netdev/HAL modules to indicate that a given PHY is supported.

The default PHY choice was updated to select only allow the supported
PHYs.

These symbols are to be used inside the HAL for various purposes
including conditional compilation for additional fields of the
'ieee802154_phy_conf_t' structure when PHY specific configurations
are added.

Signed-off-by: Jean Pierre Dudey <me@jeandudey.tech>